+++ src/osaka_sqlx/mod.rs
@@ -1,8 +1,6 @@
 pub mod booru_blacklisted_tag;
 pub mod booru_setting;
 
-use sqlx::types::BigDecimal;
-
 pub enum Fall {
     Through,
 }
@@ -11,4 +9,4 @@ pub struct ID<T> {
     pub id: T,
 }
 
-pub type BigID = ID<BigDecimal>;
+pub type BigID = ID<i64>;
